test: isolate Miller mutation regressions
